Move session error hierarchy from session/ to types/
memory/auto-save.ts (a domain module) imported BudgetExhaustedError
from src/session/errors.ts (a composition-layer module), the only
real layer violation in a small architectural audit pass.

Errors are throw-anywhere / catch-anywhere by nature, so the whole
hierarchy (SessionError, SessionNotReadyError, SessionClosedError,
BudgetExhaustedError, plus the SessionErrorCode discriminant) moves
to src/types/ -- the existing leaf-shaped home for shared types.
This removes the violation without weakening the typing of any
catch site.

Pure file move plus import-path updates. The public re-export from
src/session/index.ts is preserved (now pointed at ../types/errors.js)
so callers that imported via the session barrel keep working.
No API or behavior changes.
